HLFIR bridge: Phase 5a/5b/6 + 5c-A — allocatable, pointer, parametric struct dim, AoS+allocatable
Bridge passes
- new hlfir-rewrite-sequence-association: collapses Fortran §12.4.1.5
  scalar-element-as-array-actual adapter (ref<T> → ref<array<?xT>>) into
  explicit section designates.  Five variant traces: literal, constant
  arithmetic, parameter-constant, runtime-symbolic, multi-dim QE/BLAS.
- hlfir-flatten-structs extended for allocatable scalar-struct local member
  (5a), pointer struct member (5b), AoS+allocatable with uniform constant
  allocate sizes (5c-A).  5c-A emits static 2D companion A_<member>(N, M);
  kernel-internal allocate/deallocate sites become semantic no-ops.
- hlfir-rewrite-pointer-assigns: slice-target rebind path, inlined-callee
  alias collapse, dead-store sequential rebinds; loud failures for
  interleaved rebind/read, element-form target, bounds remap.

Bridge AST extraction
- buildSectionScalarAssign: collect read AccessInfo on scalar RHS.
- dispatch.cpp: peel fir.convert before matching hlfir.designate.
- extract_vars.cpp: detect both INT64_MIN and -1 unknown-extent sentinels;
  peel pointer-attr declares only when results have live uses.
- traceToDecl: walk through fir.rebox / fir.embox.

SDFG construction
- post-gen empty CFG region sweep — every region with zero blocks gets a
  single empty start state.  Generalised principle: empty CFG region is
  fine, just give it an empty start state.
- bindings: Fortran wrapper LOGICAL ↔ logical(c_bool) bridge.

Test fixes + 23 new tests covering all phases.

Sweep delta vs prior pause point: 463P / 34xF / 31F → 500P / 28xF / 0F.
+37 net passes, −6 xfailed, −31 failed, 0 regressions.

+++ tests/hlfir/count_intrinsic_test.py
@@ -52,7 +52,7 @@ end subroutine main
     sdfg = build_sdfg(src, tmp_path, name='main').build()
     rng = np.random.default_rng(0)
     n = 16
-    mask = (rng.random(n) > 0.5).astype(np.int32)
+    mask = (rng.random(n) > 0.5)
     res = np.zeros(1, dtype=np.int32)
     sdfg(mask=mask, n=n, res=res)
     assert int(res[0]) == int(mask.sum())
@@ -72,7 +72,7 @@ end subroutine main
     sdfg = build_sdfg(src, tmp_path, name='main').build()
     rng = np.random.default_rng(1)
     n, m = 6, 8
-    mask = np.asfortranarray((rng.random((n, m)) > 0.5).astype(np.int32))
+    mask = np.asfortranarray(rng.random((n, m)) > 0.5)
     res = np.zeros(1, dtype=np.int32)
     sdfg(mask=mask, n=n, m=m, res=res)
     assert int(res[0]) == int(mask.sum())
@@ -99,7 +99,7 @@ end subroutine main
     sdfg = build_sdfg(src, tmp_path, name='main').build()
     rng = np.random.default_rng(2)
     n, m = 5, 7
-    mask = np.asfortranarray((rng.random((n, m)) > 0.5).astype(np.int32))
+    mask = np.asfortranarray(rng.random((n, m)) > 0.5)
     res = np.zeros(n, dtype=np.int32)
     sdfg(mask=mask, n=n, m=m, res=res)
     np.testing.assert_array_equal(res, mask.sum(axis=1))
@@ -119,7 +119,7 @@ end subroutine main
     sdfg = build_sdfg(src, tmp_path, name='main').build()
     rng = np.random.default_rng(3)
     n, m = 4, 6
-    mask = np.asfortranarray((rng.random((n, m)) > 0.5).astype(np.int32))
+    mask = np.asfortranarray(rng.random((n, m)) > 0.5)
     res = np.zeros(m, dtype=np.int32)
     sdfg(mask=mask, n=n, m=m, res=res)
     np.testing.assert_array_equal(res, mask.sum(axis=0))
